Move test to fp designated file
