+++ tests/js/client/aql/aql-optimizer-rule-async-prefetch-noncluster.js
@@ -26,13 +26,8 @@
 // //////////////////////////////////////////////////////////////////////////////
 
 const jsunity = require("jsunity");
-const internal = require('internal');
-const db = internal.db;
-const helper = require("@arangodb/aql-helper");
-const errors = internal.errors;
-const assertQueryError = helper.assertQueryError;
-let IM = global.instanceManager;
- const {aql} = require('@arangodb');
+const db = require('internal').db;
+
 function optimizerRuleTestSuite () {
   const ruleName = "async-prefetch";
 
@@ -140,60 +135,6 @@ function optimizerRuleTestSuite () {
   };
 }
 
-function AsyncPrefetchFailure () {
-  const ruleName = "async-prefetch";
-
-  const cn = "UnitTestsCollection";
-  let col;
-
-  return {
-    setUpAll : function () {
-      col = db._create(cn);
-      let docs = [];
-      for(let i = 0; i < 10000; ++i) {
-       docs.push({val: i});
-      }
-      col.save(docs);
-    },
-    
-    tearDownAll : function () {
-      db._drop(cn);
-    },
-
-    // This failure was observed when the exectution block were not destroyed in topological order
-    // and the async task had outlived its execution block
-    testAsyncPrefetchFailure : function () {
-      IM.debugSetFailAt("AsyncPrefetch::blocksDestroyedOutOfOrder");
-      // This query is forced to have async prefetch optimization
-      const query = aql`FOR d IN ${col} FILTER d.val > 0 LET x = FAIL('failed') RETURN d`;
-      query.options = {optimizer: {rules: ["-move-filters-into-enumerate", "-move-filters-into-enumerate2", "-fuse-filters"]}};
-      const expectedNodes = [["SingletonNode", false], ["EnumerateCollectionNode", true], ["CalculationNode", true], ["FilterNode", true], ["CalculationNode", false], ["ReturnNode", false]];
-
-      const result = db._createStatement(query).explain();
-
-      const actualNodes = result.plan.nodes.map((n) => [n.type, n.isAsyncPrefetchEnabled === true ? true : false]);
-      assertEqual(expectedNodes, actualNodes, query);
-      
-      let expectPrefetchRule = expectedNodes.filter((n) => n[1]).length > 0;
-      if (expectPrefetchRule) {
-        assertNotEqual(-1, result.plan.rules.indexOf(ruleName));
-      } else {
-        assertEqual(-1, result.plan.rules.indexOf(ruleName));
-      }
-      assertQueryError(
-          errors.ERROR_QUERY_FAIL_CALLED.code,
-          query,
-          query.bindParams,
-          query.options
-      );
-
-      IM.debugRemoveFailAt("AsyncPrefetch::blocksDestroyedOutOfOrder");
-    },
-  };
-}
-
-
 jsunity.run(optimizerRuleTestSuite);
-jsunity.run(AsyncPrefetchFailure);
 
 return jsunity.done();
